+++ src/JQLCallback.cs
@@ -4,5 +4,5 @@ using System.Text;
 
 namespace Ejdb2
 {
-    public delegate uint JQLCallback(uint id, string json);
+    public delegate uint JQLCallback(long id, string json);
 }
